Let migration service add levels to startup
...plus other cleanup.

+++ app/services/one_off/subscription_stage_one_migration_service.rb
@@ -2,6 +2,7 @@ module OneOff
   # This service migrates data for stage 1 of the transition from 6-month program to continuous intake SaaS model.
   class SubscriptionStageOneMigrationService
     def execute
+      add_level_to_startups
       update_chores_and_sessions
       level_startups
       upgrade_legacy_startups
@@ -9,6 +10,17 @@ module OneOff
     end
 
     private
+
+    def add_level_to_startups
+      first_level = Level.find_by(number: 1)
+      raise 'Could not find a level number 1. Have all levels been created properly?' if first_level.blank?
+      startups_without_level = Startup.includes(:level).where(levels: { id: nil })
+
+      if startups_without_level.any?
+        Rails.logger.info "There are #{startups_without_level.count} startups without level. Setting them to level 1..."
+        startups_without_level.update(level: first_level)
+      end
+    end
 
     # Remove chores and sessions from target groups.
     def update_chores_and_sessions
@@ -25,7 +37,12 @@ module OneOff
     end
 
     def remove_empty_target_groups
-      TargetGroup.includes(:targets).where(targets: { id: nil }).destroy_all
+      empty_target_groups = TargetGroup.includes(:targets).where(targets: { id: nil })
+
+      if empty_target_groups.any?
+        Rails.logger.info "Removing #{empty_target_groups.count} empty targets groups..."
+        empty_target_groups.destroy_all
+      end
     end
 
     # Move startups to their appropriate level.
@@ -34,6 +51,7 @@ module OneOff
 
       Startup.where(batch: current_batch).each do |startup|
         while Startups::LevelUpEligibilityService.new(startup).eligible?
+          Rails.logger.debug "Levelling up ##{startup.id} #{startup.name} to level #{startup.level.number + 1}..."
           Startups::LevelUpService.new(startup).execute
         end
       end
@@ -41,17 +59,22 @@ module OneOff
 
     # Set iteration to 2 for startups from earlier batches.
     def upgrade_legacy_startups
-      Rails.logger.info 'Upgrading legacy startups...'
+      legacy_startups = Startup.where.not(batch: current_batch)
 
-      Startup.where.not(batch: current_batch).update(iteration: 2)
+      if legacy_startups.any?
+        Rails.logger.info "Upgrading #{legacy_startups.count} legacy startups..."
+        legacy_startups.update(iteration: 2)
+      end
     end
 
     # Remove targets that don't belong to a group.
     def remove_stale_targets
-      Rails.logger.info 'Removing stale targets...'
-
       targets_without_group = Target.includes(:target_group).where(target_groups: { id: nil }, session_at: nil, chore: false)
-      targets_without_group.destroy_all
+
+      if targets_without_group.any?
+        Rails.logger.info "Removing #{targets_without_group.count} stale targets..."
+        targets_without_group.destroy_all
+      end
     end
 
     def current_batch
